Domain-Refactoring: Categories
We unify the definition of categories and move them to the `commons`
module. We add `enumeratum` to the dependencies of `commons`.
