+++ modules/api/src/main/scala/cards/nine/api/CirceCoders.scala
@@ -1,7 +1,7 @@
 package cards.nine.api
 
 import cards.nine.api.messages.rankings
-import cards.nine.services.free.domain.Category
+import cards.nine.domain.application.Category
 import enumeratum.{ Circe ⇒ CirceEnum }
 import io.circe.generic.semiauto._
 import io.circe.{ Decoder, Encoder, ObjectEncoder }
