+++ modules/api/src/main/scala/cards/nine/api/NineCardsApi.scala
@@ -12,11 +12,11 @@ import cards.nine.api.messages.SharedCollectionMessages._
 import cards.nine.api.messages.UserMessages._
 import cards.nine.api.utils.SprayMarshallers._
 import cards.nine.api.utils.SprayMatchers._
+import cards.nine.domain.application.Category
 import cards.nine.commons.NineCardsService.Result
 import cards.nine.processes.NineCardsServices._
 import cards.nine.processes._
 import cards.nine.processes.messages.ApplicationMessages.GetAppsInfoResponse
-import cards.nine.services.free.domain.Category
 import cards.nine.services.free.domain.rankings._
 import spray.http.StatusCodes.NotFound
 import spray.routing._
