+++ modules/api/src/main/scala/cards/nine/api/messages/Rankings.scala
@@ -1,7 +1,7 @@
 package cards.nine.api.messages
 
 import cats.data.Xor
-import cards.nine.services.free.domain.Category
+import cards.nine.domain.application.Category
 
 import org.joda.time.{ DateTime, DateTimeZone }
 import org.joda.time.format.DateTimeFormat
