+++ modules/api/src/main/scala/cards/nine/api/utils/SprayMatchers.scala
@@ -1,7 +1,7 @@
 package cards.nine.api.utils
 
 import cards.nine.api.messages.PathEnumerations.PriceFilter
-import cards.nine.services.free.domain.Category
+import cards.nine.domain.application.Category
 import cards.nine.services.free.domain.rankings.{ Continent, Country }
 import enumeratum.{ Enum, EnumEntry }
 import shapeless._
