+++ modules/services/src/main/scala/cards/nine/services/free/interpreter/analytics/Converters.scala
@@ -1,6 +1,7 @@
 package cards.nine.services.free.interpreter.analytics
 
-import cards.nine.services.free.domain.{ Category, PackageName }
+import cards.nine.domain.application.Category
+import cards.nine.services.free.domain.PackageName
 import cards.nine.services.free.domain.rankings.{ GeoScope ⇒ DomainScope, _ }
 
 object Converters {
