+++ modules/services/src/main/scala/cards/nine/services/free/interpreter/ranking/Services.scala
@@ -4,8 +4,8 @@ import java.security.MessageDigest
 import java.util.UUID
 
 import cards.nine.commons.NineCardsService.Result
+import cards.nine.domain.application.{ Category, Moments }
 import cards.nine.services.free.algebra.Ranking._
-import cards.nine.services.free.domain.{ Category, Moments }
 import cards.nine.services.free.domain.rankings._
 import cards.nine.services.persistence.{ CustomComposite, Persistence }
 import cats.syntax.either._
