+++ modules/services/src/test/scala/cards/nine/services/free/domain/queries/RankingQueriesSpec.scala
@@ -1,6 +1,7 @@
 package cards.nine.services.free.domain.queries
 
-import cards.nine.services.free.domain.{ Category, PackageName }
+import cards.nine.domain.application.Category
+import cards.nine.services.free.domain.PackageName
 import cards.nine.services.free.domain.rankings._
 import cards.nine.services.persistence.DomainDatabaseContext
 import doobie.contrib.specs2.analysisspec.AnalysisSpec
