+++ modules/services/src/test/scala/cards/nine/services/free/interpreter/ranking/ServicesSpec.scala
@@ -1,6 +1,6 @@
 package cards.nine.services.free.interpreter.ranking
 
-import cards.nine.services.free.domain.Category
+import cards.nine.domain.application.Category
 import cards.nine.services.free.domain.rankings.{ CategoryRanking, Entry, GeoScope, Queries, RankedApp, Ranking, UnrankedApp }
 import cards.nine.services.persistence.{ CustomComposite, DomainDatabaseContext, NineCardsScalacheckGen }
 import org.specs2.ScalaCheck
