+++ modules/services/src/test/scala/cards/nine/services/persistence/NineCardsScalacheckGen.scala
@@ -6,8 +6,9 @@ import java.time.Instant
 import cats.Monad
 import cats.syntax.traverse._
 import cats.instances.list._
+import cards.nine.domain.application.Category
 import cards.nine.services.free.domain.rankings._
-import cards.nine.services.free.domain.{ Category, PackageName }
+import cards.nine.services.free.domain.PackageName
 import cards.nine.services.free.interpreter.collection.Services.SharedCollectionData
 import cards.nine.services.free.interpreter.user.Services.UserData
 import cards.nine.services.persistence.NineCardsGenEntities._
